interrupt: Add basic framework for interrupts
Add basic interrupt framework and implement them cpu-side, including EI,
DI, and RETI instructions. Apparently there is some timer thing that
needs to happen to delay the effect of the EI instruciton, so I
implemented that too. It is so far unclear to me if the same thing
should be happening for DI on a CGB; for a DMG it seems not. All of this
is pretty much untested.

Also add a request system and a request for the VBLANK interrupt. This
causes the boot rom to advance to the logo check and fails! Yay!