cpu: partial-borrow
Add the partial-borrow crate. This will allow us to eventually store
everything in one big struct and define interfaces between each unit and
the reset of the system.

Refactor the CPU to use partial-borrow. Introduce the CpuBus trait as
the interface between the CPU and the rest of the system. This is also
nice for tests, as they can now use their own super simple
implementation of system memory.