system(cpu): Ignore writes to cartridge ROM
For games without MBC like Tetris, the write signal is not connected to
anything. For now, ignore all writes to cartridge ROM. Once we implement
MBCs, this of course will need to be more complicated.

This fixes an issue where the score jumps really high after clearing
lines in Tetris. It turns out that Tetris actually tries to write 0x01
to 0x2000, even though it doesn't do anything on real hardware. This
resulted in modifying some function that calculates the score increase
when lines are cleared (whaddaya know..). I found an old Reddit thread
suggesting that maybe this is leftover from using MBC1 during the
development process.