refactor: Decentralize I/O register storage
Previously, I/O registers were stored contiguously in the MemoryData
struct. This inhibited modularization of code. This changes registers to
be located with the associated functionality. Overall, this means we can
use smaller bus traits (as most registers will be co-located with
related code). This should hopefully make things more modular and unit
testable.

While refactoring I noticed a bug in the PPU that was causing it to
render too fast. The old implementation skipped most of the cycles for
the Transfer mode, and caused other modes to be 1 cycle too long.

Another benefit of this refactor is that it removes register
duplication. The PPU no longer needs special initialization code to
synchronise the external registers with its internal copy: all registers
are internal and are the source of truth.

DMA now uses its own specialized read_8 routine. I think this is closer
to being accurate to real hardware, but I'm not 100% sure.

I also had to remove the addr_to_ref[_mut] functions and associated
macros. This means some extra code duplication between the read and
write paths, but they ended up not sharing much anyways.

I've also replaced the MappedReg enum with constants in the new reg
module. This makes the registers easier to use in match statements, as
unfortunately constant expressions are not available in match arm
patterns. I've used a macro to minify this a little, which lets the
comment table display nicely.

Interrupts are now much nicer to work with. IE and IF are now co-located
with the interrupt request and handling code. The joypad similarly
encapsulates its state better.