cart: MBC3 + RTC
MBC3 was pretty simple itself.

The RTC for MBC3 was more interesting. The current implementation is a
total guess (pandocs are unclear on a lot of the specifics). I don't
have fancy tools to experiment on real hardware.

The current implementation assumes the following:
 - There are two registers that store the time, a counter and a latched
   register.
 - The counter register counts every time the internal clock ticks
   (ANDed with the halt bit)
 - Latching happens on the rising edge of setting the latch signal. When
   latched, the current time is copied from the counter to the latched
   register.
 - The latch signal is taken to be the LSB of the latch signal
   register.
 - Reading the time reads from the latched register.
 - Writing the time writes directly to the counter register and doesn't
   re-latch the latch register automatically.
 - Unused bits in the flags register are ignored. Day counter MSB is
   latched along with everything else. Day carry only updates during
   latch.

RTC makes heavy use of std::time. I'll have to figure out how to deal
with this when I eventually compile for WASM. There's probably some
drop-in-replacement crate or something.