system: handle per-frame execution
Refactor such that CgbSystem is responsible for coordinating all of the
system components for a single frame, instead of handling that logic in
the fronted code. This also ensures that we sleep for the same amount of
time that has passed in the emulation, even while the LCD is being
toggled on or off.