audio: initial implementation
All channels appear to be minimally working. The APU code is kind of a
mess and is crammed into one file. Now that I have it minimally working,
I have some ideas for how it could be refactored in the future.

The audio presentation code uses cpal to play audio in a hopefully
cross-platform way. Audio is resampled from the native 2 MHz down to the
default sample rate of the audio output device. I'm not sure, maybe 4
MHz audio is needed to be cycle-accurate, but that's not what I'm going
for here.

Audio is generated synchronously with emulation and is resampled and
then put into a concurrent queue, which the audio thread pulls samples
from. If this queue ever fills up or empties out completely, it can
cause audio artifacts.

To compensate for slightly differing generation and consumption rates,
the queue length is used to dynamically adjust the resampling rate. This
keeps the queue length at a manageable size. The code for this could be
better tuned in the future. This bends the pitch of the output audio
very slightly, but the pitch bend is capped at a 3 cent radius, so it
should not be audible to most normal humans.