apu: Refactor
Initial implementation for the APU was jammed into a single file. Now
that I had a better idea of what the structure needed to look like, I
decided to refactor this and implement some abstractions to share code.

Final code actually ended up longer than the original, but it is now
sharing the concept of a "length timer" and "period divider" across
channels, which should be more modular.

The period sweep functionality could probably be refactored a bit more,
but that's a project for another day.

The period dividers should now be more accurate. Before many period
times for many channels were off-by-one. Now a period of zero is
correctly interpreted as the lowest frequency.

I also noticed a bug in the period sweep implementation during the
refactor, just a typo. That is fixed in the new version, which fixes the
Nidorino jump sound effects at the beginning of Pokemon Red.