Separate core code from frontend code
This changes this project to use a cargo workspace and two separate
packages: core and frontend. The core crate contains the core emulation
code, with the system abstracted away. Core has few dependencies, and
does not provide a user interface. Instead, it outputs to a frame buffer
and an audio callback which can be provided by users. The frontend crate
depends on core and provides the UI.

This should help separate emulation from user/UI concerns as a build out
the UI more. This also is useful if anyone ever wants to use the core in
a headless tool for something, perhaps as a testing tool or to aid in a
disassembler. Although, one should probably prefer a more
accuracy-focused emulator for such a task. Separating out core also
allows for multiple UI implementations, should that become necessary. It
also may be useful to test the core separately from the frontend.

This was pretty simple to do, as most emulation files were already
pretty independent. The one exception was joypad.rs, which contained
some `use winit::...` statements. This was not an issue; the offending
code was easily moved to the frontend package.