frontend/file-dialog: Add basic file dialog, fix web audio
File dialog uses egui-file on desktop and a custom browser-based dialog
on web. It's still just an MVP and very crunchy. But we can now play
games on the web version!

Saving no longer works; will fix that in a later commit.

Audio is now also barely working on web, but is still very buggy/hacky.
It will need some more work.