frontend: Add side panel
Remove the old test window and move the "Load ROM" button to a side
panel. The panel can be closed by clicking outside of it, and opened by
mousing over the left edge of the application.

Add a control cheatsheet at the bottom of the side panel. This was way
more difficult than it should have been.