frontend: Introduce ROM file path text box and reset button
On desktop, right now you are forced to use the file chooser to pick a
ROM file to load (unless you specify it on the command line). It is
common to provide some kind of user-editable file path on desktop
platforms to manually type in the path.

On many desktop platforms, this turns out to me more difficult than you
would think at first, as file paths might contain invalid unicode. To
solve this problem, I've added a new crate (egui-osstr) that provides an
interface to edit an osstr. It tracks which positions in the file path
have invalid unicode and treats them as opaque blobs, but allows the
user to edit the parts of the path that contain valid unicode. Right now
this only works on unix with possibly-invalid utf8. Windows support will
have to work differently because I hear they use utf16 paths over there,
but it shouldn't be too hard to implement when the time comes. I just
don't have a good way to test it right now.

On web, things have to work a bit differently, as the browser will only
let the application access files that it believes the user has
"allowed". The egui interface is opaque to the browser, so our interface
cannot provide a user-editable path and must rely on the file chooser.
Because of this, I've split the RomChooser implementation for desktop
and web.

I've also added a "Reset" button that just re-reads whatever ROM file is
currently selected and restarts the emulator. This button is also useful
on desktop to start the emulator without opening the file chooser.